control music & volume with hand detection model
In this python project I used hand detectection model to control music change previous and next, pause and play. and also control system volume  using cv2 and mediapipe and pyautogui libraries .when system camera detect open palm it will automatically play songs or any media you are using in background and when we again show open palm it will pause that media to play next media we  show 2 open finger and if we want to play previous media we use single finger . thumb up increases the system volume and the fist decreases the system volume.